ci/requirements.txt
click
requests
gitpython
